SPOP-432 Power Up populate fix

+++ Assets/Scripts/Base/Model/LevelData.cs
@@ -57,7 +57,7 @@ namespace Model
         }
 
         virtual public int ShotCount { get { return shotCount; } }
-        virtual public float[] PowerUpFills { get { return powerUpFills; } }
+        virtual public int[] PowerUpFills { get { return powerUpFills; } }
         virtual public IEnumerable<BubbleData> Bubbles { get { return bubbles; } }
         virtual public int[] StarValues { get { return starValues; } }
         virtual public BubbleQueueDefinition Queue { get { return queue; } }
@@ -69,7 +69,7 @@ namespace Model
         protected int shotCount;
 
         [SerializeField]
-        protected float[] powerUpFills;
+        protected int[] powerUpFills;
 
         [SerializeField]
         protected BubbleData[] bubbles;
