+++ Assets/Scripts/LevelEditor/Properties/LevelProperties.cs
@@ -8,7 +8,7 @@ namespace LevelEditor.Properties
         [PropertyDisplayAttribute(
             new string[] { "Blue PowerUp Fills", "Yellow PowerUp Fills", "Red PowerUp Fills", "Green PowerUp Fills", "Pink PowerUp Fills", "Purple PowerUp Fills" },
             new string[] { "3484d8", "ea9115", "d22d2b", "11952a", "e746a1", "891fc4" })]
-        public float[] PowerUpFills { get; set; }
+        public int[] PowerUpFills { get; set; }
 
         [PropertyDisplayAttribute(
             new string[] { "Star 1 Score", "Star 2 Score", "Star 3 Score" })]
@@ -17,7 +17,7 @@ namespace LevelEditor.Properties
         public LevelProperties()
         {
             StarValues = new int[] { 100, 500, 1000 };
-            PowerUpFills = new float[6];
+            PowerUpFills = new int[6];
         }
 
         public void FromLevelData(LevelData data)
